services/products/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
asyncio==4.0.0
asyncpg==0.31.0
bcrypt==4.0.1
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.122.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.11
packaging==25.0
passlib==1.7.4
prometheus_client==0.23.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rignore==0.7.6
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.0
uvicorn==0.38.0
uvloop==0.22.1
websockets==15.0.1
